+++ src/database/migrations/0_0_0_011_create_error_logs_table.php
@@ -26,7 +26,7 @@ class CreateErrorLogsTable extends Migration
             $table->integer('count')->unsigned()->default(0);
             $table->text('time_history')->nullable();
             $table->timestamps();
-            $table->bigInteger('site_id')->unsigned()->default(1);
+            $table->integer('site_id')->unsigned()->default(1);
 
             $table->primary(['id', 'site_id']);
         });
