+++ src/database/migrations/0_0_0_013_create_terms_table.php
@@ -14,17 +14,17 @@ class CreateTermsTable extends Migration
     public function up()
     {
         Schema::create('terms', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
             $table->string('name');
             $table->string('slug');
             $table->string('description')->nullable();
             $table->enum('type', ['tag', 'category'])->default('category');
             $table->enum('group', ['post', 'product'])->default('post');
-            $table->bigInteger('parent_id')->unsigned()->nullable()->index();
+            $table->integer('parent_id')->unsigned()->nullable()->index();
             $table->enum('status', ['active', 'inactive'])->default('inactive');
             $table->timestamps();
             $table->softDeletes();
-            $table->bigInteger('site_id')->unsigned()->default(1);
+            $table->integer('site_id')->unsigned()->default(1);
         });
     }
